Removed robots.txt. Wrong project to add this to!
robots.txt is located directly under the root domain. Revive & Restore
is mounted as a sub-directory which does not need it's own robots.txt.
Instead, this project and djlongnow will need to be aware of each other
and the robots.txt will be included in the djlongnow project as that is
what picks up when requesting the root domain.